+++ archive-nb/build-archive-nb.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# Copyright 2026, University of Maryland, All Rights Reserved
+
 # Copy archive-specific notebooks
 set -eu