+++ env-core/build-r.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# Copyright 2026, University of Maryland, All Rights Reserved
+
 # Add R kernel
 set -eux
 set -o pipefail