+++ changes.md
@@ -4,6 +4,7 @@
 - Update fornax-labextension to 0.1.22: Change the label Introduction to Tutorials.
 - Add s3archive script to help efficiently archive data to ~/s3-storage.
 - Bug fixes and enhancements:
+  - Firefly now uses IRSA's cloud backend.
 
 ## 26.0814
 - Hotfix for the missing /opt/support-data symlink.