+++ changes.md
@@ -1,6 +1,7 @@
 # Latest Changes
 
 ## dev
+- Upgrade the default python to 3.13.
 - Update fornax-labextension to 0.1.22: Change the label Introduction to Tutorials.
 - Add s3archive script to help efficiently archive data to ~/s3-storage.
 - Add jupyter-ai extension.